Changes for login and account management and templates for vendor inventory management.

+++ TestAmazeLocal/src/com/iu/amazelocal/utils/AppConstants.java
@@ -2,5 +2,5 @@ package com.iu.amazelocal.utils;
 
 public class AppConstants {
 	public static final Long USERIDSEQ=100000L;
-	public static final int LOGINIDSEQ=200000;
+	public static final Long LOGINIDSEQ=200000L;
 }
